Simplify key/pk field with KeyValue wrapper.
The pk property now returns a KeyValue instance, which adds a unicode
method which outputs the urlsafe value, and a(nother) pk property which
returns the value; this triggers the correct behaviour in inline
formsets' _construct_form method. KeyValue hashes/compares equal to its
key value.

This now allows the removal of a lot of the customised form code, as
well as the confusing model `id` property and the overridden admin
submit_line template.